@@ -5,29 +5,36 @@ import (
 	"compress/gzip"
 	"context"
 	"crypto/tls"
-	"encoding/base64"
-	"encoding/json"
 	"fmt"
 	"io"
-	"io/ioutil"
 	"net"
 	"net/http"
 	"net/http/httptest"
 	"net/url"
 	"os"
+	"regexp"
+	"strconv"
 	"strings"
 	"sync"
 	"testing"
 	"time"
 
-	"github.com/contentsquare/chproxy/cache"
-
 	"github.com/alicebob/miniredis/v2"
+	"github.com/contentsquare/chproxy/cache"
 	"github.com/contentsquare/chproxy/config"
 	"github.com/contentsquare/chproxy/log"
+	"github.com/prometheus/client_golang/prometheus"
+	"github.com/stretchr/testify/assert"
+
+	"github.com/contentsquare/chproxy/global/types"
 )
 
-var testDir = "./temp-test-data"
+var (
+	testDir   = "./temp-test-data"
+	labels    = prometheus.Labels{}
+	redisPort = types.RedisPort
+	chPort    = types.ClickHousePort
+)
 
 func TestMain(m *testing.M) {
 	log.SuppressOutput(true)
@@ -50,7 +57,7 @@ func TestServe(t *testing.T) {
 		name     string
 		file     string
 		testFn   func(t *testing.T)
-		listenFn func() (net.Listener, chan struct{})
+		listenFn func() (*http.Server, chan struct{})
 	}{
 		{
 			"https request",
@@ -91,13 +98,16 @@ func TestServe(t *testing.T) {
 				if resp.StatusCode != http.StatusOK {
 					t.Fatalf("unexpected status code: %d; expected: %d", resp.StatusCode, http.StatusOK)
 				}
-				checkHttpResponse(t, resp, expectedOkResp)
+				checkResponse(t, resp.Body, expectedOkResp)
+				checkHeader(t, resp, "X-Cache", "MISS")
 
 				// check cached response
+				credHash, _ := calcCredentialHash("default", "qwerty")
 				key := &cache.Key{
-					Query:          []byte(q),
-					AcceptEncoding: "gzip",
-					Version:        cache.Version,
+					Query:              []byte(q),
+					AcceptEncoding:     "gzip",
+					Version:            cache.Version,
+					UserCredentialHash: credHash,
 				}
 				path := fmt.Sprintf("%s/cache/%s", testDir, key.String())
 				if _, err := os.Stat(path); err != nil {
@@ -111,11 +121,125 @@ func TestServe(t *testing.T) {
 				if err != nil {
 					t.Fatalf("unexpected error while getting response from cache: %s", err)
 				}
-				err = RespondWithData(rw, cachedData.Data, cachedData.ContentMetadata, cachedData.Ttl, 200)
+				err = RespondWithData(rw, cachedData.Data, cachedData.ContentMetadata, cachedData.Ttl, XCacheHit, 200, labels)
 				if err != nil {
 					t.Fatalf("unexpected error while getting response from cache: %s", err)
 				}
 				checkResponse(t, rw.Body, expectedOkResp)
+				checkHeader(t, rw.Result(), "X-Cache", XCacheHit)
+			},
+			startTLS,
+		},
+		{
+			"https cache max payload size",
+			"testdata/https.cache.max-payload-size.yml",
+			func(t *testing.T) {
+				q := "SELECT MaxPayloadSize"
+				req, err := http.NewRequest("GET", "https://127.0.0.1:8443?query="+url.QueryEscape(q), nil)
+				checkErr(t, err)
+				req.SetBasicAuth("default", "qwerty")
+				req.Close = true
+
+				resp, err := tlsClient.Do(req)
+				checkErr(t, err)
+				if resp.StatusCode != http.StatusOK {
+					t.Fatalf("unexpected status code: %d; expected: %d", resp.StatusCode, http.StatusOK)
+				}
+
+				checkResponse(t, resp.Body, expectedOkResp)
+				checkHeader(t, resp, "X-Cache", XCacheNA)
+
+				key := &cache.Key{
+					Query:          []byte(q),
+					AcceptEncoding: "gzip",
+					Version:        cache.Version,
+				}
+
+				cc := proxy.caches["https_cache_max_payload_size"]
+				cachedData, err := cc.Get(key)
+
+				if cachedData != nil || err == nil {
+					t.Fatal("response bigger than maxPayloadSize should not be cached")
+				}
+
+				resp.Body.Close()
+			},
+			startTLS,
+		},
+		{
+			"https cache max payload size not reached",
+			"testdata/https.cache.max-payload-size-not-reached.yml",
+			func(t *testing.T) {
+				q := "SELECT MaxPayloadSize"
+				req, err := http.NewRequest("GET", "https://127.0.0.1:8443?query="+url.QueryEscape(q), nil)
+				checkErr(t, err)
+				req.SetBasicAuth("default", "qwerty")
+				req.Close = true
+
+				resp, err := tlsClient.Do(req)
+				checkErr(t, err)
+				if resp.StatusCode != http.StatusOK {
+					t.Fatalf("unexpected status code: %d; expected: %d", resp.StatusCode, http.StatusOK)
+				}
+
+				checkResponse(t, resp.Body, expectedOkResp)
+
+				credHash, _ := calcCredentialHash("default", "qwerty")
+				key := &cache.Key{
+					Query:              []byte(q),
+					AcceptEncoding:     "gzip",
+					Version:            cache.Version,
+					UserCredentialHash: credHash,
+				}
+
+				rw := httptest.NewRecorder()
+
+				cc := proxy.caches["https_cache_max_payload_size"]
+				cachedData, err := cc.Get(key)
+
+				if err != nil {
+					t.Fatalf("unexpected error while getting response from cache: %s", err)
+				}
+
+				err = RespondWithData(rw, cachedData.Data, cachedData.ContentMetadata, cachedData.Ttl, XCacheHit, 200, labels)
+				if err != nil {
+					t.Fatalf("unexpected error while getting response from cache: %s", err)
+				}
+				checkResponse(t, rw.Body, expectedOkResp)
+
+				cachedData.Data.Close()
+				resp.Body.Close()
+			},
+			startTLS,
+		},
+		{
+			"https request body is not empty",
+			"testdata/https.yml",
+			func(t *testing.T) {
+				query := "SELECT SleepTimeout"
+				buf := bytes.NewBufferString(query)
+				req, err := http.NewRequest("POST", "https://127.0.0.1:8443", buf)
+				checkErr(t, err)
+				req.SetBasicAuth("default", "qwerty")
+				req.Close = true
+
+				resp, err := tlsClient.Do(req)
+				checkErr(t, err)
+				if resp.StatusCode != http.StatusGatewayTimeout {
+					t.Fatalf("unexpected status code: %d; expected: %d", resp.StatusCode, http.StatusGatewayTimeout)
+				}
+
+				bodyBytes, err := io.ReadAll(resp.Body)
+				if err != nil {
+					t.Fatalf("error while reading body from response; err: %q", err)
+				}
+
+				b := string(bodyBytes)
+				if !strings.Contains(b, query) {
+					t.Fatalf("expected request body: %q; got: %q", query, b)
+				}
+
+				resp.Body.Close()
 			},
 			startTLS,
 		},
@@ -139,11 +263,14 @@ func TestServe(t *testing.T) {
 				resp.Body.Close()
 
 				// check cached response
+				credHash, _ := calcCredentialHash("default", "qwerty")
 				key := &cache.Key{
-					Query:          []byte(expectedQuery),
-					AcceptEncoding: "gzip",
-					Version:        cache.Version,
+					Query:              []byte(expectedQuery),
+					AcceptEncoding:     "gzip",
+					Version:            cache.Version,
+					UserCredentialHash: credHash,
 				}
+
 				path := fmt.Sprintf("%s/cache/%s", testDir, key.String())
 				if _, err := os.Stat(path); err != nil {
 					t.Fatalf("err while getting file %q info: %s", path, err)
@@ -157,7 +284,7 @@ func TestServe(t *testing.T) {
 					t.Fatalf("unexpected error while writing reposnse from cache: %s", err)
 				}
 
-				err = RespondWithData(rw, cachedData.Data, cachedData.ContentMetadata, cachedData.Ttl, 200)
+				err = RespondWithData(rw, cachedData.Data, cachedData.ContentMetadata, cachedData.Ttl, XCacheMiss, 200, labels)
 				if err != nil {
 					t.Fatalf("unexpected error while getting response from cache: %s", err)
 				}
@@ -178,16 +305,19 @@ func TestServe(t *testing.T) {
 				req.SetBasicAuth("default", "qwerty")
 				resp, err := tlsClient.Do(req)
 				checkErr(t, err)
-				if resp.StatusCode != http.StatusTeapot {
-					t.Fatalf("unexpected status code: %d; expected: %d", resp.StatusCode, http.StatusTeapot)
+				if resp.StatusCode != http.StatusInternalServerError {
+					t.Fatalf("unexpected status code: %d; expected: %d", resp.StatusCode,
+						http.StatusInternalServerError)
 				}
 				resp.Body.Close()
 
 				// check cached response
+				credHash, _ := calcCredentialHash("default", "qwerty")
 				key := &cache.Key{
-					Query:          []byte(q),
-					AcceptEncoding: "gzip",
-					Version:        cache.Version,
+					Query:              []byte(q),
+					AcceptEncoding:     "gzip",
+					Version:            cache.Version,
+					UserCredentialHash: credHash,
 				}
 				path := fmt.Sprintf("%s/cache/%s", testDir, key.String())
 				if _, err := os.Stat(path); !os.IsNotExist(err) {
@@ -349,27 +479,30 @@ func TestServe(t *testing.T) {
 				q := "SELECT redis_cache_mate"
 				req, err := http.NewRequest("GET", "http://127.0.0.1:9090?query="+url.QueryEscape(q), nil)
 				checkErr(t, err)
-
-				resp := httpRequest(t, req, http.StatusOK)
-				checkHttpResponse(t, resp, expectedOkResp)
-				resp2 := httpRequest(t, req, http.StatusOK)
-				checkHttpResponse(t, resp2, expectedOkResp)
 				keys := redisClient.Keys()
-				if len(keys) != 1 {
+				// redis should be empty before the test
+				if len(keys) != 0 {
+					t.Fatalf("unexpected amount of keys in redis: %v", len(keys))
+				}
+
+				resp, err := httpRequest(t, req, http.StatusOK)
+				checkErr(t, err)
+				checkResponse(t, resp.Body, expectedOkResp)
+				resp2, err := httpRequest(t, req, http.StatusOK)
+				checkErr(t, err)
+				checkResponse(t, resp2.Body, expectedOkResp)
+				keys = redisClient.Keys()
+				if len(keys) != 2 { // expected 2 because there is a record stored for transaction and a cache item
 					t.Fatalf("unexpected amount of keys in redis: %v", len(keys))
 				}
 
 				// check cached response
+				credHash, _ := calcCredentialHash("default", "")
 				key := &cache.Key{
-					Query:          []byte(q),
-					AcceptEncoding: "gzip",
-					Version:        cache.Version,
-				}
-				str, err := redisClient.Get(key.String())
-				checkErr(t, err)
-
-				if !strings.Contains(str, base64.StdEncoding.EncodeToString([]byte("Ok."))) || !strings.Contains(str, "text/plain") || !strings.Contains(str, "charset=utf-8") {
-					t.Fatalf("result from cache query is wrong: %s", str)
+					Query:              []byte(q),
+					AcceptEncoding:     "gzip",
+					Version:            cache.Version,
+					UserCredentialHash: credHash,
 				}
 
 				duration := redisClient.TTL(key.String())
@@ -385,46 +518,25 @@ func TestServe(t *testing.T) {
 			func(t *testing.T) {
 				redisClient.FlushAll()
 				q := "SELECT 1 FORMAT TabSeparatedWithNamesAndTypes"
-				req, err := http.NewRequest("GET", "http://127.0.0.1:9090?query="+url.QueryEscape(q), nil)
-				checkErr(t, err)
-
-				resp := httpRequest(t, req, http.StatusOK)
-				checkHttpResponse(t, resp, string(bytesWithInvalidUTFPairs))
-				resp2 := httpRequest(t, req, http.StatusOK)
-				// if we do not use base64 to encode/decode the cached payload, EOF error will be thrown here.
-				checkHttpResponse(t, resp2, string(bytesWithInvalidUTFPairs))
 				keys := redisClient.Keys()
-				if len(keys) != 1 {
+				// redis should be empty before the test
+				if len(keys) != 0 {
 					t.Fatalf("unexpected amount of keys in redis: %v", len(keys))
 				}
 
-				// check cached response
-				key := &cache.Key{
-					Query:          []byte(q),
-					AcceptEncoding: "gzip",
-					Version:        cache.Version,
-				}
-				str, err := redisClient.Get(key.String())
+				req, err := http.NewRequest("GET", "http://127.0.0.1:9090?query="+url.QueryEscape(q), nil)
 				checkErr(t, err)
 
-				type redisCachePayload struct {
-					Length   int64  `json:"l"`
-					Type     string `json:"t"`
-					Encoding string `json:"enc"`
-					Payload  string `json:"payload"`
-				}
-
-				var unMarshaledPayload redisCachePayload
-				err = json.Unmarshal([]byte(str), &unMarshaledPayload)
+				resp, err := httpRequest(t, req, http.StatusOK)
 				checkErr(t, err)
-				if unMarshaledPayload.Payload != base64.StdEncoding.EncodeToString(bytesWithInvalidUTFPairs) {
-					t.Fatalf("result from cache query is wrong: %s", str)
-				}
-				decoded, err := base64.StdEncoding.DecodeString(unMarshaledPayload.Payload)
+				checkResponse(t, resp.Body, string(bytesWithInvalidUTFPairs))
+				resp2, err := httpRequest(t, req, http.StatusOK)
 				checkErr(t, err)
-
-				if unMarshaledPayload.Length != int64(len(decoded)) {
-					t.Fatalf("the declared length %d and actual length %d is not same", unMarshaledPayload.Length, len(decoded))
+				// if we do not use base64 to encode/decode the cached payload, EOF error will be thrown here.
+				checkResponse(t, resp2.Body, string(bytesWithInvalidUTFPairs))
+				keys = redisClient.Keys()
+				if len(keys) != 2 { // 2 because there is a record stored for transaction, and a cache item
+					t.Fatalf("unexpected amount of keys in redis: %v", len(keys))
 				}
 			},
 			startHTTP,
@@ -444,9 +556,10 @@ func TestServe(t *testing.T) {
 				req.Header.Set("Content-Encoding", "gzip")
 				resp, err := http.DefaultClient.Do(req)
 				checkErr(t, err)
-				body, _ := ioutil.ReadAll(resp.Body)
+				body, _ := io.ReadAll(resp.Body)
 				if resp.StatusCode != http.StatusOK {
-					t.Fatalf("unexpected status code: %d; expected: %d; body: %s", resp.StatusCode, http.StatusOK, string(body))
+					t.Fatalf("unexpected status code: %d; expected: %d; body: %s", resp.StatusCode, http.StatusOK,
+						string(body))
 				}
 				resp.Body.Close()
 			},
@@ -523,6 +636,16 @@ func TestServe(t *testing.T) {
 			startHTTP,
 		},
 		{
+			"http metrics namespace",
+			"testdata/http.metrics.namespace.yml",
+			func(t *testing.T) {
+				resp := httpGet(t, "http://127.0.0.1:9090/metrics", http.StatusOK)
+				assert.GreaterOrEqual(t, len(getStringFromResponse(t, resp.Body)), 10000)
+				resp.Body.Close()
+			},
+			startHTTP,
+		},
+		{
 			"http user networks",
 			"testdata/http.user.networks.yml",
 			func(t *testing.T) {
@@ -545,20 +668,91 @@ func TestServe(t *testing.T) {
 			startHTTP,
 		},
 		{
-			"http shared cache",
+			"http shared cache for them user",
 			"testdata/http.shared.cache.yml",
 			func(t *testing.T) {
 				// actually we can check that cache-file is shared via metrics
 				// but it needs additional library for doing this
 				// so it would be just files amount check
-				cacheDir := "temp-test-data/shared_cache"
+				cacheDir := "temp-test-data/shared_cache1"
 				checkFilesCount(t, cacheDir, 0)
-				httpGet(t, "http://127.0.0.1:9090?query=SELECT&user=user1", http.StatusOK)
+				httpGet(t, "http://127.0.0.1:9090?query=SELECT&user=user1_test1", http.StatusOK)
 				checkFilesCount(t, cacheDir, 1)
-				httpGet(t, "http://127.0.0.1:9090?query=SELECT&user=user2", http.StatusOK)
+				httpGet(t, "http://127.0.0.1:9090?query=SELECT&user=user1_test1", http.StatusOK)
 				// request from different user expected to be served with already cached,
 				// so count of files should be the same
 				checkFilesCount(t, cacheDir, 1)
+			},
+			startHTTP,
+		},
+		{
+			"http not share cache for same user with different pwd",
+			"testdata/http.shared.cache.yml",
+			func(t *testing.T) {
+				// because of the wildcarded feature that delegate the authentication to clickouse
+				// we can't afford to return a cached of the same user without reaching clickhouse
+				// if the pwd is not the same than the one used for the cached query
+				cacheDir := "temp-test-data/shared_cache2"
+				checkFilesCount(t, cacheDir, 0)
+				httpGet(t, "http://127.0.0.1:9090?query=SELECT&user=user1_test2&password=toto", http.StatusOK)
+				checkFilesCount(t, cacheDir, 1)
+				httpGet(t, "http://127.0.0.1:9090?query=SELECT&user=user2_test2&password=titi", http.StatusOK)
+				checkFilesCount(t, cacheDir, 2)
+			},
+			startHTTP,
+		},
+		{
+			"http not share cache for different user",
+			"testdata/http.shared.cache.yml",
+			func(t *testing.T) {
+				cacheDir := "temp-test-data/shared_cache3"
+				checkFilesCount(t, cacheDir, 0)
+				httpGet(t, "http://127.0.0.1:9090?query=SELECT&user=user1_test3", http.StatusOK)
+				checkFilesCount(t, cacheDir, 1)
+				httpGet(t, "http://127.0.0.1:9090?query=SELECT&user=user2_test3", http.StatusOK)
+				checkFilesCount(t, cacheDir, 2)
+			},
+			startHTTP,
+		},
+		{
+			"http share cache for same user with different pwd",
+			"testdata/http.shared.cache.yml",
+			func(t *testing.T) {
+				// because of the wildcarded feature that delegate the authentication to clickouse
+				// we can't afford to return a cached of the same user without reaching clickhouse
+				// if the pwd is not the same than the one used for the cached query
+				cacheDir := "temp-test-data/shared_cache4"
+				checkFilesCount(t, cacheDir, 0)
+				httpGet(t, "http://127.0.0.1:9090?query=SELECT&user=user1_test4&password=toto", http.StatusOK)
+				checkFilesCount(t, cacheDir, 1)
+				httpGet(t, "http://127.0.0.1:9090?query=SELECT&user=user2_test4&password=titi", http.StatusOK)
+				checkFilesCount(t, cacheDir, 1)
+			},
+			startHTTP,
+		},
+		{
+			"http share cache for different user",
+			"testdata/http.shared.cache.yml",
+			func(t *testing.T) {
+				cacheDir := "temp-test-data/shared_cache5"
+				checkFilesCount(t, cacheDir, 0)
+				httpGet(t, "http://127.0.0.1:9090?query=SELECT&user=user1_test5", http.StatusOK)
+				checkFilesCount(t, cacheDir, 1)
+				httpGet(t, "http://127.0.0.1:9090?query=SELECT&user=user2_test5", http.StatusOK)
+				checkFilesCount(t, cacheDir, 1)
+			},
+			startHTTP,
+		},
+		{
+			"http not share cache for different user using wildcarded feature",
+			"testdata/http.shared.cache.yml",
+			func(t *testing.T) {
+				cacheDir := "temp-test-data/shared_cache6"
+				checkFilesCount(t, cacheDir, 0)
+				httpGet(t, "http://127.0.0.1:9090?query=SELECT&user=user1_test6", http.StatusOK)
+				checkFilesCount(t, cacheDir, 1)
+				httpGet(t, "http://127.0.0.1:9090?query=SELECT&user=user2_test6", http.StatusOK)
+				checkFilesCount(t, cacheDir, 2)
 			},
 			startHTTP,
 		},
@@ -583,9 +777,14 @@ func TestServe(t *testing.T) {
 				req = req.WithContext(ctx)
 				_, err = http.DefaultClient.Do(req)
 				expErr := "context deadline exceeded"
+				if err == nil {
+					t.Fatal("expected deadline error")
+				}
+
 				if !strings.Contains(err.Error(), "context deadline exceeded") {
 					t.Fatalf("unexpected error: %s; expected: %s", err, expErr)
 				}
+				// verifies if query isn't cached and if kill query was launched.
 				select {
 				case <-fakeCHState.syncCH:
 					// wait while chproxy will detect that request was canceled and will drop temp file
@@ -597,6 +796,68 @@ func TestServe(t *testing.T) {
 			},
 			startHTTP,
 		},
+		{
+			"http concurrent transaction from cache",
+			"testdata/http.concurrent.transaction.yml",
+			func(t *testing.T) {
+
+				// max_exec_time = 300 ms, grace_time = 160 ms (> max_exec_time/2)
+				// scenario: 1st query completes before grace_time elapsed. 2nd query is served from cache.
+
+				q := "SELECT SLEEP100"
+				executeTwoConcurrentRequests(t, q, http.StatusOK, http.StatusOK, "success mate", "success mate")
+			},
+			startHTTP,
+		},
+		{
+			"http concurrent transaction failure scenario",
+			"testdata/http.concurrent.transaction.yml",
+			func(t *testing.T) {
+				// max_exec_time = 300 ms, grace_time = 160 ms (> max_exec_time/2)
+				// scenario: 1st query fails before grace_time elapsed. 2nd query fails as well.
+
+				q := "SELECT ERROR"
+				executeTwoConcurrentRequests(t, q, http.StatusInternalServerError, http.StatusInternalServerError,
+					"DB::Exception\n", "[concurrent query failed] DB::Exception\n")
+			},
+			startHTTP,
+		},
+		{
+			"http concurrent transaction failure scenario - transaction completed, not failed - query is recoverable",
+			"testdata/http.concurrent.transaction.yml",
+			func(t *testing.T) {
+				// max_exec_time = 300 ms, grace_time = 160 ms (> max_exec_time/2)
+				// scenario: 1st query fails before grace_time elapsed. 2nd query fails as well.
+
+				q := "SELECT RECOVERABLE-ERROR"
+				executeTwoConcurrentRequests(t, q, http.StatusServiceUnavailable, http.StatusServiceUnavailable,
+					"DB::Unavailable\n", "DB::Unavailable\n")
+			},
+			startHTTP,
+		},
+		{
+			"http request with default proxy headers",
+			"testdata/https.proxy-enabled.yml",
+			func(t *testing.T) {
+				query := "SELECT 1"
+				req, err := http.NewRequest("GET", "https://127.0.0.1:8443?query="+url.QueryEscape(query), nil)
+				checkErr(t, err)
+				req.Close = true
+				req.SetBasicAuth("default", "qwerty")
+				req.Header.Add("X-Forwarded-For", "10.0.0.1")
+
+				resp, err := tlsClient.Do(req)
+				checkErr(t, err)
+				defer resp.Body.Close()
+
+				if resp.StatusCode != http.StatusOK {
+					t.Fatalf("unexpected status code: %d; expected: %d", resp.StatusCode, http.StatusOK)
+				}
+
+				checkResponse(t, resp.Body, expectedOkResp)
+			},
+			startTLS,
+		},
 	}
 
 	// Wait until CHServer starts.
@@ -605,7 +866,7 @@ func TestServe(t *testing.T) {
 	startTime := time.Now()
 	i := 0
 	for i < 10 {
-		if _, err := http.Get("http://127.0.0.1:8124/"); err == nil {
+		if _, err := http.Get(fmt.Sprintf("http://127.0.0.1:%s/", chPort)); err == nil {
 			break
 		}
 		time.Sleep(10 * time.Millisecond)
@@ -615,13 +876,15 @@ func TestServe(t *testing.T) {
 		t.Fatalf("CHServer didn't start in %s", time.Since(startTime))
 	}
 
+	cfg := &config.Config{}
+	registerMetrics(cfg)
 	for _, tc := range testCases {
 		t.Run(tc.name, func(t *testing.T) {
 			*configFile = tc.file
-			ln, done := tc.listenFn()
+			s, done := tc.listenFn()
 			defer func() {
-				if err := ln.Close(); err != nil {
-					t.Fatalf("unexpected error while closing listener: %s", err)
+				if err := s.Shutdown(context.Background()); err != nil {
+					t.Fatalf("unexpected error while closing server: %s", err)
 				}
 				<-done
 			}()
@@ -648,27 +911,13 @@ func TestServe(t *testing.T) {
 	}
 }
 
-func checkHttpResponse(t *testing.T, resp *http.Response, expectedResp string) {
-	buf := new(bytes.Buffer)
-	_, err := buf.ReadFrom(resp.Body)
-	if err != nil {
-		t.Fatalf("unexpected error while reading a response body: %v", err)
-	}
-	responseBodyStr := buf.String()
-	resp.Body.Close()
-
-	if responseBodyStr != expectedResp {
-		t.Fatalf("got: %q; expected: %q", responseBodyStr, expectedResp)
-	}
-}
-
 func checkFilesCount(t *testing.T, dir string, expectedLen int) {
-	files, err := ioutil.ReadDir(dir)
+	files, err := os.ReadDir(dir)
 	if err != nil {
 		t.Fatalf("error while reading dir %q: %s", dir, err)
 	}
 	if expectedLen != len(files) {
-		t.Fatalf("expected %d files in cache dir; got: %d", expectedLen, len(files))
+		t.Fatalf("expected %d files in cache dir %s; got: %d", expectedLen, dir, len(files))
 	}
 }
 
@@ -676,7 +925,7 @@ var tlsClient = &http.Client{Transport: &http.Transport{
 	TLSClientConfig: &tls.Config{InsecureSkipVerify: true},
 }}
 
-func startTLS() (net.Listener, chan struct{}) {
+func startTLS() (*http.Server, chan struct{}) {
 	cfg, err := loadConfig()
 	if err != nil {
 		panic(fmt.Sprintf("error while loading config: %s", err))
@@ -689,17 +938,21 @@ func startTLS() (net.Listener, chan struct{}) {
 	if err != nil {
 		panic(fmt.Sprintf("cannot listen for %q: %s", cfg.Server.HTTPS.ListenAddr, err))
 	}
-	tlsCfg := newTLSConfig(cfg.Server.HTTPS)
+	tlsCfg, err := cfg.Server.HTTPS.TLS.BuildTLSConfig(autocertManager)
+	if err != nil {
+		panic(fmt.Sprintf("cannot build TLS config: %s", err))
+	}
 	tln := tls.NewListener(ln, tlsCfg)
 	h := http.HandlerFunc(serveHTTP)
+	s := newServer(tln, h, config.TimeoutCfg{})
 	go func() {
-		listenAndServe(tln, h, config.TimeoutCfg{})
+		s.Serve(tln)
 		close(done)
 	}()
-	return tln, done
+	return s, done
 }
 
-func startHTTP() (net.Listener, chan struct{}) {
+func startHTTP() (*http.Server, chan struct{}) {
 	cfg, err := loadConfig()
 	if err != nil {
 		panic(fmt.Sprintf("error while loading config: %s", err))
@@ -713,17 +966,18 @@ func startHTTP() (net.Listener, chan struct{}) {
 		panic(fmt.Sprintf("cannot listen for %q: %s", cfg.Server.HTTP.ListenAddr, err))
 	}
 	h := http.HandlerFunc(serveHTTP)
+	s := newServer(ln, h, config.TimeoutCfg{})
 	go func() {
-		listenAndServe(ln, h, config.TimeoutCfg{})
+		s.Serve(ln)
 		close(done)
 	}()
-	return ln, done
+	return s, done
 }
 
 // TODO randomise port for each instance of the mock
 func startRedis(t *testing.T) *miniredis.Miniredis {
 	redis := miniredis.NewMiniRedis()
-	if err := redis.StartAddr("127.0.0.1:6379"); err != nil {
+	if err := redis.StartAddr("127.0.0.1:" + redisPort); err != nil {
 		t.Fatalf("Failed to create redis server: %s", err)
 	}
 	return redis
@@ -731,8 +985,10 @@ func startRedis(t *testing.T) *miniredis.Miniredis {
 
 // TODO randomise port for each instance of the mock
 func startCHServer() {
-	http.ListenAndServe(":8124", http.HandlerFunc(fakeCHHandler))
+	http.ListenAndServe(":"+chPort, http.HandlerFunc(fakeCHHandler))
 }
+
+var patt = regexp.MustCompile(`(\d+)$`)
 
 func fakeCHHandler(w http.ResponseWriter, r *http.Request) {
 	query, err := getFullQuery(r)
@@ -746,29 +1002,137 @@ func fakeCHHandler(w http.ResponseWriter, r *http.Request) {
 		fmt.Fprint(w, "got empty query for non-GET request")
 		return
 	}
-	switch string(query) {
-	case "SELECT ERROR":
-		w.WriteHeader(http.StatusTeapot)
-		fmt.Fprint(w, "DB::Exception\n")
-	case "SELECT SLEEP":
-		w.WriteHeader(http.StatusOK)
-		fmt.Fprint(w, "foo")
+	defer func() {
 		if f, ok := w.(http.Flusher); ok {
 			f.Flush()
 		}
-
+	}()
+	q := string(query)
+	switch {
+	case q == "SELECT ERROR":
+		w.WriteHeader(http.StatusInternalServerError)
+		fmt.Fprint(w, "DB::Exception\n")
+	case q == "SELECT RECOVERABLE-ERROR":
+		println("called clickhouse recoverable")
+		w.WriteHeader(http.StatusServiceUnavailable)
+		fmt.Fprint(w, "DB::Unavailable\n")
+	case q == "SELECT SLEEP":
+		w.WriteHeader(http.StatusOK)
+		fmt.Fprint(w, "foo")
 		fakeCHState.sleep()
-
 		fmt.Fprint(w, "bar")
-	case "SELECT 1 FORMAT TabSeparatedWithNamesAndTypes":
+
+	case strings.Contains(q, "SELECT SLEEP"):
+		numberStrings := patt.FindAllStringSubmatch(q, -1)
+		if len(numberStrings) == 0 {
+			panic("couldn't extract number from sleep call to clickhouse mock")
+		}
+		nr, err := strconv.Atoi(numberStrings[0][1])
+		if err != nil {
+			panic(err)
+		}
+
+		duration, err := time.ParseDuration(fmt.Sprintf("%vms", nr))
+		if err != nil {
+			panic(err)
+		}
+		time.Sleep(duration)
+		w.WriteHeader(http.StatusOK)
+		fmt.Fprint(w, "success mate") //nolint
+	case q == "SELECT 1 FORMAT TabSeparatedWithNamesAndTypes":
 		w.WriteHeader(http.StatusOK)
 		w.Write(bytesWithInvalidUTFPairs)
+	case q == "SELECT MaxPayloadSize":
+		w.WriteHeader(http.StatusOK)
+
+		// generate 10M payload size
+		b := make([]byte, 10485760)
+		w.Header().Set("Content-Type", "text/plain; charset=utf-8")
+		fmt.Fprint(w, b)
+		fmt.Fprint(w, "Ok.\n")
+	case strings.Contains(q, "SELECT SleepTimeout"):
+		w.WriteHeader(http.StatusGatewayTimeout)
+
+		bodyBytes, err := io.ReadAll(r.Body)
+		if err != nil {
+			fmt.Fprintf(w, "query: %s; error while reading body: %s", query, err)
+			return
+		}
+
+		b := string(bodyBytes)
+		// Ensure the original request body is not empty and remains unchanged
+		// after it is processed by getFullQuery.
+		if b == "" && b != q {
+			fmt.Fprintf(w, "got original req body: <%s>; escaped query: <%s>", b, q)
+			return
+		}
+
+		// execute sleep 1.5 sec
+		time.Sleep(1500 * time.Millisecond)
+		fmt.Fprint(w, b)
 	default:
 		if strings.Contains(string(query), killQueryPattern) {
 			fakeCHState.kill()
 		}
 		w.WriteHeader(http.StatusOK)
 		fmt.Fprint(w, "Ok.\n")
+	}
+}
+
+// executeTwoConcurrentRequests concurrently executes 2 requests for the same query.
+// Results are asserted according to the specified input parameters.
+func executeTwoConcurrentRequests(t *testing.T, query string, firstStatusCode, secondStatusCode int,
+	firstBody, secondBody string) {
+	u := fmt.Sprintf("http://127.0.0.1:9090?query=%s&user=concurrent_user", url.QueryEscape(query))
+
+	var wg sync.WaitGroup
+	wg.Add(2)
+	var resp1 string
+	var resp2 string
+	errs := make(chan error, 0)
+	defer close(errs)
+	errors := make([]error, 0)
+	go func() {
+		for err := range errs {
+			errors = append(errors, err)
+		}
+	}()
+	go func() {
+		defer wg.Done()
+		req, err := http.NewRequest("GET", u, nil)
+		checkErr(t, err)
+		resp, err := httpRequest(t, req, firstStatusCode)
+		if err != nil {
+			errs <- err
+			return
+		}
+		resp1 = bbToString(t, resp.Body)
+	}()
+
+	go func() {
+		defer wg.Done()
+		time.Sleep(20 * time.Millisecond)
+		req, err := http.NewRequest("GET", u, nil)
+		checkErr(t, err)
+		resp, err := httpRequest(t, req, secondStatusCode)
+		if err != nil {
+			errs <- err
+			return
+		}
+		resp2 = bbToString(t, resp.Body)
+	}()
+	wg.Wait()
+
+	if len(errors) != 0 {
+		t.Fatalf("concurrent test scenario failed due to: %v", errors)
+	}
+
+	if !strings.Contains(resp1, firstBody) {
+		t.Fatalf("concurrent test scenario: unexpected resp body: %s, expected : %s", resp1, firstBody)
+	}
+
+	if !strings.Contains(resp2, secondBody) {
+		t.Fatalf("concurrent test scenario: unexpected resp body: %s, expected : %s", resp2, secondBody)
 	}
 }
 
@@ -782,12 +1146,6 @@ type stateCH struct {
 	sync.Mutex
 	inited bool
 	syncCH chan struct{}
-}
-
-func (s *stateCH) isInited() bool {
-	s.Lock()
-	defer s.Unlock()
-	return s.inited
 }
 
 func (s *stateCH) kill() {
@@ -807,25 +1165,31 @@ func (s *stateCH) sleep() {
 }
 
 func TestNewTLSConfig(t *testing.T) {
-	cfg := config.HTTPS{
+	cfg := &config.TLS{
 		KeyFile:  "testdata/example.com.key",
 		CertFile: "testdata/example.com.cert",
 	}
 
-	tlsCfg := newTLSConfig(cfg)
+	tlsCfg, err := cfg.BuildTLSConfig(autocertManager)
+	if err != nil {
+		panic(fmt.Sprintf("cannot build TLS config: %s", err))
+	}
 	if len(tlsCfg.Certificates) < 1 {
 		t.Fatalf("expected tls certificate; got empty list")
 	}
 
 	certCachePath := fmt.Sprintf("%s/certs_dir", testDir)
-	cfg = config.HTTPS{
+	cfg = &config.TLS{
 		Autocert: config.Autocert{
 			CacheDir:     certCachePath,
 			AllowedHosts: []string{"example.com"},
 		},
 	}
 	autocertManager = newAutocertManager(cfg.Autocert)
-	tlsCfg = newTLSConfig(cfg)
+	if err != nil {
+		panic(fmt.Sprintf("cannot build TLS config: %s", err))
+	}
+	tlsCfg, _ = cfg.BuildTLSConfig(autocertManager)
 	if tlsCfg.GetCertificate == nil {
 		t.Fatalf("expected func GetCertificate be set; got nil")
 	}
@@ -853,15 +1217,19 @@ func checkErr(t *testing.T, err error) {
 	}
 }
 
-func checkResponse(t *testing.T, r io.Reader, expected string) {
+func getStringFromResponse(t *testing.T, r io.Reader) string {
 	if r == nil {
-		t.Fatal("unexpected nil reader")
+		t.Fatalf("unexpected nil reader")
 	}
-	response, err := ioutil.ReadAll(r)
+	response, err := io.ReadAll(r)
 	if err != nil {
 		t.Fatalf("unexpected err while reading: %s", err)
 	}
-	got := string(response)
+	return string(response)
+}
+
+func checkResponse(t *testing.T, r io.Reader, expected string) {
+	got := getStringFromResponse(t, r)
 	if !strings.Contains(got, expected) {
 		t.Fatalf("got: %q; expected: %q", got, expected)
 	}
@@ -878,14 +1246,26 @@ func httpGet(t *testing.T, url string, statusCode int) *http.Response {
 	return resp
 }
 
-func httpRequest(t *testing.T, request *http.Request, statusCode int) *http.Response {
+func checkHeader(t *testing.T, resp *http.Response, header string, expected string) {
+	t.Helper()
+
+	h := resp.Header
+	v := h.Get(header)
+
+	if v != expected {
+		t.Fatalf("for header: %s got: %s, expected %s", header, v, expected)
+	}
+}
+
+func httpRequest(t *testing.T, request *http.Request, statusCode int) (*http.Response, error) {
+	t.Helper()
 	client := http.Client{}
 	resp, err := client.Do(request)
 	if err != nil {
-		t.Fatalf("unexpected erorr while doing GET request: %s", err)
+		return resp, fmt.Errorf("unexpected erorr while doing GET request: %s", err)
 	}
 	if resp.StatusCode != statusCode {
-		t.Fatalf("unexpected status code: %d; expected: %d", resp.StatusCode, statusCode)
+		return resp, fmt.Errorf("unexpected status code: %d; expected: %d", resp.StatusCode, statusCode)
 	}
-	return resp
+	return resp, nil
 }
