@@ -8,90 +8,110 @@ import (
 
 	"github.com/alicebob/miniredis/v2"
 	"github.com/contentsquare/chproxy/config"
+	"github.com/stretchr/testify/assert"
 )
 
 const asyncTestDir = "./async-test-data"
 
 func TestAsyncCache_Cleanup_Of_Expired_Transactions(t *testing.T) {
 	graceTime := 100 * time.Millisecond
-	asyncCache := newAsyncTestCache(t, graceTime)
+	asyncCache := newAsyncTestCache(t, graceTime, graceTime/2)
+	defer func() {
+		asyncCache.Close()
+		os.RemoveAll(asyncTestDir)
+	}()
 
 	key := &Key{
 		Query: []byte("SELECT async cache"),
 	}
-
-	if done := asyncCache.IsDone(key); !done {
+	status, err := asyncCache.Status(key)
+	assert.NoError(t, err)
+	if !status.State.IsAbsent() {
 		t.Fatalf("unexpected behaviour: transaction isnt done while it wasnt even started")
 	}
 
-	if err := asyncCache.Register(key); err != nil {
+	if err := asyncCache.Create(key); err != nil {
 		t.Fatalf("unexpected error: %s failed to register transaction", err)
 	}
 
-	if done := asyncCache.IsDone(key); done {
+	status, err = asyncCache.Status(key)
+	assert.NoError(t, err)
+	if !status.State.IsPending() {
 		t.Fatalf("unexpected behaviour: transaction isnt finished")
 	}
 
 	time.Sleep(graceTime * 2)
 
-	if done := asyncCache.IsDone(key); !done {
+	status, err = asyncCache.Status(key)
+	assert.NoError(t, err)
+	if status.State.IsPending() {
 		t.Fatalf("unexpected behaviour: transaction grace time elapsed and yet it was still pending")
 	}
-
-	asyncCache.Close()
-	os.RemoveAll(asyncTestDir)
 }
 
 func TestAsyncCache_AwaitForConcurrentTransaction_GraceTimeWithoutTransactionCompletion(t *testing.T) {
-	graceTime := 300 * time.Millisecond
-	asyncCache := newAsyncTestCache(t, graceTime)
+	graceTime := 100 * time.Millisecond
+	asyncCache := newAsyncTestCache(t, graceTime, graceTime/2)
+
+	defer func() {
+		asyncCache.Close()
+		os.RemoveAll(asyncTestDir)
+	}()
 
 	key := &Key{
 		Query: []byte("SELECT async cache AwaitForConcurrentTransaction"),
 	}
 
-	if done := asyncCache.IsDone(key); !done {
+	status, err := asyncCache.Status(key)
+	assert.NoError(t, err)
+	if !status.State.IsAbsent() {
 		t.Fatalf("unexpected behaviour: transaction isnt done while it wasnt even started")
 	}
 
-	if err := asyncCache.Register(key); err != nil {
+	if err := asyncCache.Create(key); err != nil {
 		t.Fatalf("unexpected error: %s failed to register transaction", err)
 	}
 
-	if done := asyncCache.IsDone(key); done {
+	status, err = asyncCache.Status(key)
+	assert.NoError(t, err)
+	if !status.State.IsPending() {
 		t.Fatalf("unexpected behaviour: transaction isnt finished")
 	}
 
 	startTime := time.Now()
-	done := asyncCache.AwaitForConcurrentTransaction(key)
+	_, err = asyncCache.AwaitForConcurrentTransaction(key)
+	assert.NoError(t, err)
 	elapsedTime := time.Since(startTime)
 
 	// in order to let the cleaner swipe the transaction
-	time.Sleep(100 * time.Millisecond)
-	if done == asyncCache.IsDone(key) && done {
+	time.Sleep(150 * time.Millisecond)
+	status, err = asyncCache.Status(key)
+	assert.NoError(t, err)
+	if !status.State.IsAbsent() {
 		t.Fatalf("unexpected behaviour: transaction awaiting time elapsed %s", elapsedTime.String())
 	}
-
-	asyncCache.Close()
-	os.RemoveAll(asyncTestDir)
 }
 
 func TestAsyncCache_AwaitForConcurrentTransaction_TransactionCompletedWhileAwaiting(t *testing.T) {
 	graceTime := 300 * time.Millisecond
-	asyncCache := newAsyncTestCache(t, graceTime)
+	asyncCache := newAsyncTestCache(t, graceTime, graceTime/2)
 
+	defer func() {
+		asyncCache.Close()
+		os.RemoveAll(asyncTestDir)
+	}()
 	key := &Key{
 		Query: []byte("SELECT async cache AwaitForConcurrentTransactionCompleted"),
 	}
 
-	if err := asyncCache.Register(key); err != nil {
+	if err := asyncCache.Create(key); err != nil {
 		t.Fatalf("unexpected error: %s failed to register transaction", err)
 	}
 
 	errs := make(chan error)
 	go func() {
 		time.Sleep(graceTime / 2)
-		if err := asyncCache.Unregister(key); err != nil {
+		if err := asyncCache.Complete(key); err != nil {
 			errs <- err
 		} else {
 			errs <- nil
@@ -99,23 +119,74 @@ func TestAsyncCache_AwaitForConcurrentTransaction_TransactionCompletedWhileAwait
 	}()
 
 	startTime := time.Now()
-	done := asyncCache.AwaitForConcurrentTransaction(key)
-	elapsedTime := time.Since(startTime)
-
-	err := <-errs
+	transactionState, err := asyncCache.AwaitForConcurrentTransaction(key)
 	if err != nil {
 		t.Fatalf("unexpected error: %s failed to unregister transaction", err)
 	}
 
-	if done != asyncCache.IsDone(key) || !done || elapsedTime >= graceTime {
+	elapsedTime := time.Since(startTime)
+
+	err = <-errs
+	if err != nil {
+		t.Fatalf("unexpected error: %s failed to unregister transaction", err)
+	}
+
+	if !transactionState.State.IsCompleted() || elapsedTime >= graceTime {
+		t.Fatalf("unexpected behaviour: transaction awaiting time elapsed %s", elapsedTime.String())
+	}
+}
+
+func TestAsyncCache_AwaitForConcurrentTransaction_TransactionFailedWhileAwaiting(t *testing.T) {
+	graceTime := 300 * time.Millisecond
+	asyncCache := newAsyncTestCache(t, graceTime, graceTime/2)
+
+	defer func() {
+		asyncCache.Close()
+		os.RemoveAll(asyncTestDir)
+	}()
+
+	key := &Key{
+		Query: []byte("SELECT async cache AwaitForConcurrentTransactionCompleted"),
+	}
+
+	if err := asyncCache.Create(key); err != nil {
+		t.Fatalf("unexpected error: %s failed to register transaction", err)
+	}
+
+	failReason := "failed for fun"
+	errs := make(chan error)
+	go func() {
+		time.Sleep(graceTime / 2)
+		if err := asyncCache.Fail(key, failReason); err != nil {
+			errs <- err
+		} else {
+			errs <- nil
+		}
+	}()
+
+	startTime := time.Now()
+	transactionState, err := asyncCache.AwaitForConcurrentTransaction(key)
+	if err != nil {
+		t.Fatalf("unexpected error: %s failed to unregister transaction", err)
+	}
+
+	elapsedTime := time.Since(startTime)
+
+	err = <-errs
+	if err != nil {
+		t.Fatalf("unexpected error: %s failed to unregister transaction", err)
+	}
+
+	if !transactionState.State.IsFailed() || elapsedTime >= graceTime {
 		t.Fatalf("unexpected behaviour: transaction awaiting time elapsed %s", elapsedTime.String())
 	}
 
-	asyncCache.Close()
-	os.RemoveAll(asyncTestDir)
+	if transactionState.FailReason != failReason {
+		t.Fatalf("unexpected behaviour: transaction failed but without err reason")
+	}
 }
 
-func newAsyncTestCache(t *testing.T, graceTime time.Duration) *AsyncCache {
+func newAsyncTestCache(t *testing.T, graceTime, transactionEndedTime time.Duration) *AsyncCache {
 	t.Helper()
 	cfg := config.Cache{
 		Name: "foobar",
@@ -132,7 +203,7 @@ func newAsyncTestCache(t *testing.T, graceTime time.Duration) *AsyncCache {
 
 	asyncC := &AsyncCache{
 		Cache:               c,
-		TransactionRegistry: newInMemoryTransactionRegistry(graceTime),
+		TransactionRegistry: newInMemoryTransactionRegistry(graceTime, transactionEndedTime),
 		graceTime:           graceTime,
 	}
 	return asyncC
@@ -147,12 +218,13 @@ func TestAsyncCache_FilesystemCache_instantiation(t *testing.T) {
 			Dir:     asyncTestDir,
 			MaxSize: 8192,
 		},
-		Expire: config.Duration(time.Minute),
+		Expire:         config.Duration(time.Minute),
+		MaxPayloadSize: config.ByteSize(100000000),
 	}
 	if err := os.RemoveAll(testDirAsync); err != nil {
 		log.Fatalf("cannot remove %q: %s", testDirAsync, err)
 	}
-	_, err := NewAsyncCache(fileSystemCfg)
+	_, err := NewAsyncCache(fileSystemCfg, 1*time.Second)
 	if err != nil {
 		t.Fatalf("could not instanciate filsystem async cache because of the following error: %s", err)
 	}
@@ -168,7 +240,7 @@ func TestAsyncCache_FilesystemCache_wrong_instantiation(t *testing.T) {
 		},
 		Expire: config.Duration(time.Minute),
 	}
-	_, err := NewAsyncCache(fileSystemCfg)
+	_, err := NewAsyncCache(fileSystemCfg, 1*time.Second)
 	if err == nil {
 		t.Fatalf("the instanciate of filsystem async cache should have crashed")
 	}
@@ -182,12 +254,48 @@ func TestAsyncCache_RedisCache_instantiation(t *testing.T) {
 		Redis: config.RedisCacheConfig{
 			Addresses: []string{s.Addr()},
 		},
-		Expire: config.Duration(cacheTTL),
+		Expire:         config.Duration(cacheTTL),
+		MaxPayloadSize: config.ByteSize(100000000),
 	}
 
-	_, err := NewAsyncCache(redisCfg)
+	_, err := NewAsyncCache(redisCfg, 1*time.Second)
 	if err != nil {
 		t.Fatalf("could not instanciate redis async cache because of the following error: %s", err)
+	}
+}
+
+func TestAsyncCache_RedisCache_TLS(t *testing.T) {
+	cfg := config.TLS{
+		CertFile:           "../testdata/example.com.cert",
+		KeyFile:            "../testdata/example.com.key",
+		InsecureSkipVerify: true,
+	}
+
+	tlsConfig, err := cfg.BuildTLSConfig(nil)
+	if err != nil {
+		t.Fatalf("could not build tls config: %s", err)
+	}
+	s := miniredis.NewMiniRedis()
+	if err := s.StartTLS(tlsConfig); err != nil {
+		t.Fatalf("could not start miniredis: %s", err.Error())
+		// not reached
+	}
+	t.Cleanup(s.Close)
+
+	var redisCfg = config.Cache{
+		Name: "test",
+		Mode: "redis",
+		Redis: config.RedisCacheConfig{
+			TLS:       cfg,
+			Addresses: []string{s.Addr()},
+		},
+		Expire:         config.Duration(cacheTTL),
+		MaxPayloadSize: config.ByteSize(100000000),
+	}
+
+	_, err = NewAsyncCache(redisCfg, 1*time.Second)
+	if err != nil {
+		t.Fatalf("could not instanciate redis async cache because of the following error: %s", err.Error())
 	}
 }
 
@@ -201,13 +309,13 @@ func TestAsyncCache_RedisCache_wrong_instantiation(t *testing.T) {
 		},
 	}
 
-	_, err := NewAsyncCache(redisCfg)
+	_, err := NewAsyncCache(redisCfg, 1*time.Second)
 	if err == nil {
 		t.Fatalf("the redis instanciation should have crashed")
 	}
 }
 
-func TestAsyncCache_Unkown_instantiation(t *testing.T) {
+func TestAsyncCache_Unknown_instantiation(t *testing.T) {
 	var redisCfg = config.Cache{
 		Name:   "test",
 		Mode:   "Unkown Mode",
@@ -215,7 +323,7 @@ func TestAsyncCache_Unkown_instantiation(t *testing.T) {
 		Expire: config.Duration(cacheTTL),
 	}
 
-	_, err := NewAsyncCache(redisCfg)
+	_, err := NewAsyncCache(redisCfg, 1*time.Second)
 	if err == nil {
 		t.Fatalf("The instanciation should have crash")
 	}
