@@ -5,7 +5,7 @@ import (
 	"fmt"
 	"io"
 
-	"github.com/DataDog/zstd"
+	"github.com/klauspost/compress/zstd"
 	"github.com/pierrec/lz4"
 )
 
@@ -74,7 +74,19 @@ func (r *Reader) readNextBlock() error {
 	}
 
 	// Decompress block
+	if err := r.decompressBlock(block, compressionType, decompressedSize); err != nil {
+		return err
+	}
+
+	return nil
+}
+
+func (r *Reader) decompressBlock(block []byte, compressionType byte, decompressedSize uint32) error {
+	var err error
+
 	r.data = make([]byte, decompressedSize)
+	var decoder, _ = zstd.NewReader(nil)
+
 	switch compressionType {
 	case noneType:
 		r.data = block
@@ -83,13 +95,15 @@ func (r *Reader) readNextBlock() error {
 			return fmt.Errorf("cannot decompress lz4 block: %w", err)
 		}
 	case zstdType:
-		r.data, err = zstd.Decompress(r.data, block)
+		r.data = r.data[:0] // Wipe the slice but keep allocated memory
+		r.data, err = decoder.DecodeAll(block, r.data)
 		if err != nil {
 			return fmt.Errorf("cannot decompress zstd block: %w", err)
 		}
 	default:
 		return fmt.Errorf("unknown compressionType: %X", compressionType)
 	}
+
 	return nil
 }
 
